+++ libasynql/cli/fx.php
@@ -1,3 +1,4 @@
+#!/usr/bin/env php
 <?php
 
 /*
@@ -24,7 +25,14 @@ use poggit\libasynql\generic\GenericStatementFileParseException;
 use poggit\libasynql\generic\GenericStatementFileParser;
 use poggit\libasynql\GenericStatement;
 
-require_once __DIR__ . "/../cli-autoload.php";
+if(Phar::running()) {
+	// virion cli mode
+	require_once __DIR__ . "/../cli-autoload.php";
+} else {
+	// composer bin mode
+	require_once __DIR__ . "/../../../../autoload.php";
+	array_splice($argv, 1, 0, ["fx"]);
+}
 
 function constToCamel(string $prefix, string $const) : string{
 	$camel = "";