+++ libasynql/src/poggit/libasynql/base/QueryRecvQueue.php
@@ -22,35 +22,42 @@ declare(strict_types=1);
 
 namespace poggit\libasynql\base;
 
+use pmmp\thread\ThreadSafe;
+use pmmp\thread\ThreadSafeArray;
 use poggit\libasynql\SqlError;
 use poggit\libasynql\SqlResult;
-use Threaded;
 use function is_string;
 use function serialize;
 use function unserialize;
 
-class QueryRecvQueue extends Threaded{
+class QueryRecvQueue extends ThreadSafe{
 	private int $availableThreads = 0;
+
+	private ThreadSafeArray $queue;
+
+	public function __construct(){
+		$this->queue = new ThreadSafeArray();
+	}
 
 	/**
 	 * @param SqlResult[] $results
 	 */
 	public function publishResult(int $queryId, array $results) : void{
 		$this->synchronized(function() use ($queryId, $results) : void{
-			$this[] = serialize([$queryId, $results]);
+			$this->queue[] = serialize([$queryId, $results]);
 			$this->notify();
 		});
 	}
 
 	public function publishError(int $queryId, SqlError $error) : void{
 		$this->synchronized(function() use ($error, $queryId) : void{
-			$this[] = serialize([$queryId, $error]);
+			$this->queue[] = serialize([$queryId, $error]);
 			$this->notify();
 		});
 	}
 
 	public function fetchResults(&$queryId, &$results) : bool{
-		$row = $this->shift();
+		$row = $this->queue->shift();
 		if(is_string($row)){
 			[$queryId, $results] = unserialize($row, ["allowed_classes" => true]);
 			return true;
@@ -59,25 +66,32 @@ class QueryRecvQueue extends Threaded{
 	}
 
 	/**
-	 * @param SqlError|SqlResults[]|null $results
+	 * @param SqlError|SqlResult[]|null $results
 	 */
-	public function waitForResults(?int &$queryId, SqlError|array|null &$results) : bool{
-		return $this->synchronized(function() use (&$queryId, &$results) : bool{
-			while($this->count() === 0 && $this->availableThreads > 0){
-				$this->wait();
+	public function fetchAllResults(): array{
+		return $this->synchronized(function(): array{
+			$ret = [];
+			while($this->fetchResults($queryId, $results)){
+				$ret[] = [$queryId, $results];
 			}
-			return $this->fetchResults($queryId, $results);
+			return $ret;
 		});
 	}
 
-	public function addAvailableThread() : void{
-		$this->synchronized(fn() => ++$this->availableThreads);
-	}
-
-	public function removeAvailableThread() : void{
-		$this->synchronized(function() : void{
-			--$this->availableThreads;
-			$this->notify();
+	/**
+	 * @return list<array{int, SqlError|SqlResults[]|null}>
+	 */
+	public function waitForResults(int $expectedResults): array{
+		return $this->synchronized(function() use ($expectedResults) : array{
+			$ret = [];
+			while(count($ret) < $expectedResults){
+				if(!$this->fetchResults($queryId, $results)){
+					$this->wait();
+					continue;
+				}
+				$ret[] = [$queryId, $results];
+			}
+			return $ret;
 		});
 	}
 }