+++ libasynql/src/poggit/libasynql/base/SqlSlaveThread.php
@@ -22,22 +22,18 @@ declare(strict_types=1);
 
 namespace poggit\libasynql\base;
 
-use ClassLoader;
 use InvalidArgumentException;
+use pmmp\thread\Thread as NativeThread;
 use pocketmine\Server;
-use pocketmine\snooze\SleeperNotifier;
+use pocketmine\snooze\SleeperHandlerEntry;
 use pocketmine\thread\Thread;
 use poggit\libasynql\libasynql;
 use poggit\libasynql\SqlError;
 use poggit\libasynql\SqlResult;
 use poggit\libasynql\SqlThread;
-use const PTHREADS_INHERIT_CONSTANTS;
-use const PTHREADS_INHERIT_INI;
-use function assert;
 
 abstract class SqlSlaveThread extends Thread implements SqlThread{
-	/** @var SleeperNotifier */
-	private $notifier;
+	private SleeperHandlerEntry $sleeperEntry;
 
 	private static $nextSlaveNumber = 0;
 
@@ -48,13 +44,12 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 	protected $connError;
 	protected $busy = false;
 
-	protected function __construct(SleeperNotifier $notifier, QuerySendQueue $bufferSend = null, QueryRecvQueue $bufferRecv = null){
-		$this->notifier = $notifier;
+	protected function __construct(SleeperHandlerEntry $entry, QuerySendQueue $bufferSend = null, QueryRecvQueue $bufferRecv = null){
+		$this->sleeperEntry = $entry;
 
 		$this->slaveNumber = self::$nextSlaveNumber++;
 		$this->bufferSend = $bufferSend ?? new QuerySendQueue();
 		$this->bufferRecv = $bufferRecv ?? new QueryRecvQueue();
-		$this->bufferRecv->addAvailableThread();
 
 		if(!libasynql::isPackaged()){
 			/** @noinspection PhpUndefinedMethodInspection */
@@ -63,7 +58,7 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 			$cl = Server::getInstance()->getPluginManager()->getPlugin("DEVirion")->getVirionClassLoader();
 			$this->setClassLoaders([Server::getInstance()->getLoader(), $cl]);
 		}
-		$this->start(PTHREADS_INHERIT_INI | PTHREADS_INHERIT_CONSTANTS);
+		$this->start(NativeThread::INHERIT_INI);
 	}
 
 	protected function onRun() : void{
@@ -71,14 +66,14 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 		$this->connCreated = true;
 		$this->connError = $error;
 
+		$notifier = $this->sleeperEntry->createNotifier();
+
 		if($error !== null){
 			return;
 		}
 
 		while(true){
-			$this->bufferRecv->removeAvailableThread();
 			$row = $this->bufferSend->fetchQuery();
-			$this->bufferRecv->addAvailableThread();
 			if(!is_string($row)){
 				break;
 			}
@@ -87,7 +82,7 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 
 			try{
 				$results = [];
-				foreach($queries as $index => $query) {
+				foreach($queries as $index => $query){
 					$results[] = $this->executeQuery($resource, $modes[$index], $query, $params[$index]);
 				}
 				$this->bufferRecv->publishResult($queryId, $results);
@@ -95,17 +90,16 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 				$this->bufferRecv->publishError($queryId, $error);
 			}
 
-			$this->notifier->wakeupSleeper();
+			$notifier->wakeupSleeper();
 			$this->busy = false;
 		}
-		$this->bufferRecv->removeAvailableThread();
 		$this->close($resource);
 	}
 
 	/**
 	 * @return bool
 	 */
-	public function isBusy(): bool {
+	public function isBusy() : bool{
 		return $this->busy;
 	}
 
@@ -124,8 +118,13 @@ abstract class SqlSlaveThread extends Thread implements SqlThread{
 		$this->bufferSend->scheduleQuery($queryId, $modes, $queries, $params);
 	}
 
-	public function readResults(array &$callbacks) : void{
-		while($this->bufferRecv->waitForResults($queryId, $results)){
+	public function readResults(array &$callbacks, ?int $expectedResults) : void{
+		if($expectedResults === null){
+			$resultsList = $this->bufferRecv->fetchAllResults();
+		}else{
+			$resultsList = $this->bufferRecv->waitForResults($expectedResults);
+		}
+		foreach($resultsList as [$queryId, $results]){
 			if(!isset($callbacks[$queryId])){
 				throw new InvalidArgumentException("Missing handler for query #$queryId");
 			}