+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 
-## [4.1.1](https://github.com/poggit/libasynql/compare/v4.1.1...v4.1.2)
+## [4.2.0](https://github.com/poggit/libasynql/compare/v4.1.6...v4.2.0)
+- Support for PocketMine API 5
+
+## [4.1.6](https://github.com/poggit/libasynql/compare/v4.1.5...v4.1.6)
+- Set `def` and `fx` as composer bin for running from vendor
+
+## [4.1.5](https://github.com/poggit/libasynql/compare/v4.1.4...v4.1.5)
+- Move `libasynql::detectPackaged()` call into `libasynql::create()`
+
+### Added
+- Virion 3.0 support
+
+## [4.1.4](https://github.com/poggit/libasynql/compare/v4.1.3...v4.1.4)
+
+### Added
+- Virion 3.0 support
+
+## [4.1.3](https://github.com/poggit/libasynql/compare/v4.1.2...v4.1.3)
+
+### Fixed
+- Fixed queries not executing until the next sleeper wakeup
+
+## [4.1.2](https://github.com/poggit/libasynql/compare/v4.1.1...v4.1.2)
 
 ### Added
 - Indicate query name in the "the variable is not used anywhere" error message